refactor: migrate to one file per database model
