+++ src/models/__init__.py
@@ -1,0 +1,11 @@
+# to enable from models import *
+
+from src.models.user import User # type: ignore
+from src.models.project import Project # type: ignore
+from src.models.opportunity import Opportunity # type: ignore
+from src.models.objective import Objective # type: ignore
+from src.models.graph import Graph # type: ignore
+from src.models.decision import Decision # type: ignore
+from src.models.probability import Probability # type: ignore
+from src.models.node import Node # type: ignore
+from src.models.edge import Edge # type: ignore
